# hiring-2024-IOP-transporte/api-definition/src/main/resources/itx-iop_tech-transport-hiring-openapi.yaml
openapi: 3.0.3
info:
  title: ITX IOP TRANSPORT hiring sample API
  description: |-
    ITX IOP TRANSPORT hiring sample API
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local server
tags:
  - name: Shipments
    description: Shipment operation endpoints
paths:
  /shipment:
    post:
      tags:
        - Shipments
      summary: Endpoint to load a new shipment
      description: Load a shipment into the system to be processed later
      operationId: loadShipment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShipmentInput'
      responses:
        '200':
          description: Successful operation
        '400':
          description: Invalid request
  /shipments:
    get:
      tags:
        - Shipments
      summary: Get list of current shipments
      description: Retrieve all shipments with their routing information if available
      operationId: getShipments
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Shipment'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
        - Shipments
      summary: Endpoint to purge shipments just processed (status DISCARDED or PLANNED)
      description: |-
        Endpoint to purge shipments just processed (status DISCARDED or PLANNED). PENDING shipments will not be deleted
      operationId: purgeShipments
      responses:
        '200':
          description: Successful operation
        '404':
          $ref: '#/components/responses/NotFound'
  /shipments/process:
    post:
      tags:
        - Shipments
      summary: Endpoint to process pending shipments
      description: |-
        Dijkstra algorithm to process pending shipments and calculate the best route
      operationId: processPendingShipments
      responses:
        '200':
          description: Successful operation
  /shipment/{shipmentId}:
    get:
      tags:
        - Shipments
      summary: Get global information about a concrete shipment
      description: |-
        Retrieve information related to a shipment using the shipment identifier
        Collects data about the truck details, location stops and loaded packages
      operationId: getShipment
      parameters:
        - in: path
          name: shipmentId
          required: true
          schema:
            type: string
            format: uuid
          description: Shipment identifier (UUID)
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shipment'
            application/xml:
              schema:
                $ref: '#/components/schemas/Shipment'
            text/csv:
              schema:
                type: string
        '404':
          $ref: '#/components/responses/NotFound'
components:
  responses:
    NotFound:
      description: The specified resource was not found
  schemas:
    ShipmentInput:
      type: object
      description: Payload used to ingest a shipment
      properties:
        shipmentId:
          type: string
          format: uuid
          description: Shipment identifier (UUID)
          example: 123e4567-e89b-12d3-a456-426614174000
        originCityCode:
          type: string
          minLength: 3
          maxLength: 3
          description: Origin city identifier (3-char unique identifier)
          example: SVQ
        destinationCityCode:
          type: string
          minLength: 3
          maxLength: 3
          description: Destination city identifier (3-char unique identifier)
          example: MAD
        departureDate:
          type: string
          format: date-time
          description: Departure date and time
          example: 2026-01-15T10:00:00Z
        expectedArrivalDate:
          type: string
          format: date-time
          description: Expected arrival date and time
          example: 2026-01-16T10:00:00Z
      required:
        - shipmentId
        - originCityCode
        - destinationCityCode
    Shipment:
      type: object
      description: Shipment aggregated information
      properties:
        shipmentId:
          type: string
          format: uuid
          description: Shipment identifier (UUID)
        originCity:
          $ref: '#/components/schemas/City'
        destinationCity:
          $ref: '#/components/schemas/City'
        departureDate:
          type: string
          format: date-time
          description: Departure date
        expectedArrivalDate:
          type: string
          format: date-time
          description: Expected arrival date
        status:
          type: string
          enum: [PENDING, DISCARDED, PLANNED]
          description: Shipment status
        routePlan:
          type: array
          items:
            $ref: '#/components/schemas/Route'
          description: Planned route for the shipment
      required:
        - shipmentId
        - originCity
        - destinationCity
        - status
    City:
      type: object
      properties:
          code: 
            type: string
            description: City identifier (3-char unique udentifier)
            minLength: 3
            maxLength: 3
            example: MAD
          name: 
            type: string
            description: City full name
            minLength: 1
            maxLength: 9999
            example: Madrid
          handlingCost:
            type: number
            format: double
            description: Handling cost
            example: 10.0
      required:
          - code
          - name
    Route:
      type: object
      properties:
        id:
          type: string
          description: Shipment identifier (UUID)
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        origin:
          $ref: '#/components/schemas/City'
        destination:
          $ref: '#/components/schemas/City'
        transportType:
          enum: [TRUCK, TRAIN, SEA, AIR]
          description: Transport type
          example: TRUCK
        cost:
          type: string
          format: dur-time
          description: Route time Cost (duration in hours)
          example: T12H
        handlingCost:
          type: number
          format: double
          description: Handling cost
          example: 10.0
      required:
        - code
        - name
    Error:
      type: object
      properties:
        info:
          type: string
          description: Info about the error
          example: [invalid operation service not ready]
      required: 
        - info