# hiring-2024-IOP-transporte/api-definition/target/generated-sources/openapi/src/gen/java/main/com/itxiop/transport/infrastructure/apirest/ShipmentsApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.18.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.itxiop.transport.infrastructure.apirest;

import com.itxiop.transport.infrastructure.apirest.model.Shipment;
import com.itxiop.transport.infrastructure.apirest.model.ShipmentInput;
import java.util.UUID;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-01-16T08:56:22.982580100+01:00[Europe/Madrid]", comments = "Generator version: 7.18.0")
@Validated
@Tag(name = "Shipments", description = "Shipment operation endpoints")
public interface ShipmentsApi {

    String PATH_GET_SHIPMENT = "/shipment/{shipmentId}";
    /**
     * GET /shipment/{shipmentId} : Get global information about a concrete shipment
     * Retrieve information related to a shipment using the shipment identifier Collects data about the truck details, location stops and loaded packages
     *
     * @param shipmentId Shipment identifier (UUID) (required)
     * @return Successful operation (status code 200)
     *         or The specified resource was not found (status code 404)
     */
    @Operation(
        operationId = "getShipment",
        summary = "Get global information about a concrete shipment",
        description = "Retrieve information related to a shipment using the shipment identifier Collects data about the truck details, location stops and loaded packages",
        tags = { "Shipments" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Successful operation", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Shipment.class)),
                @Content(mediaType = "application/xml", schema = @Schema(implementation = Shipment.class)),
                @Content(mediaType = "text/csv", schema = @Schema(implementation = Shipment.class))
            }),
            @ApiResponse(responseCode = "404", description = "The specified resource was not found")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = ShipmentsApi.PATH_GET_SHIPMENT,
        produces = { "application/json", "application/xml", "text/csv" }
    )
    ResponseEntity<Shipment> getShipment(
        @NotNull @Parameter(name = "shipmentId", description = "Shipment identifier (UUID)", required = true, in = ParameterIn.PATH) @PathVariable("shipmentId") UUID shipmentId
    );


    String PATH_GET_SHIPMENTS = "/shipments";
    /**
     * GET /shipments : Get list of current shipments
     * Retrieve all shipments with their routing information if available
     *
     * @return Successful operation (status code 200)
     *         or The specified resource was not found (status code 404)
     */
    @Operation(
        operationId = "getShipments",
        summary = "Get list of current shipments",
        description = "Retrieve all shipments with their routing information if available",
        tags = { "Shipments" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Successful operation", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Shipment.class)))
            }),
            @ApiResponse(responseCode = "404", description = "The specified resource was not found")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = ShipmentsApi.PATH_GET_SHIPMENTS,
        produces = { "application/json" }
    )
    ResponseEntity<List<Shipment>> getShipments(
        
    );


    String PATH_LOAD_SHIPMENT = "/shipment";
    /**
     * POST /shipment : Endpoint to load a new shipment
     * Load a shipment into the system to be processed later
     *
     * @param shipmentInput  (required)
     * @return Successful operation (status code 200)
     *         or Invalid request (status code 400)
     */
    @Operation(
        operationId = "loadShipment",
        summary = "Endpoint to load a new shipment",
        description = "Load a shipment into the system to be processed later",
        tags = { "Shipments" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Successful operation"),
            @ApiResponse(responseCode = "400", description = "Invalid request")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = ShipmentsApi.PATH_LOAD_SHIPMENT,
        consumes = { "application/json" }
    )
    ResponseEntity<Void> loadShipment(
        @Parameter(name = "ShipmentInput", description = "", required = true) @Valid @RequestBody ShipmentInput shipmentInput
    );


    String PATH_PROCESS_PENDING_SHIPMENTS = "/shipments/process";
    /**
     * POST /shipments/process : Endpoint to process pending shipments
     * Dijkstra algorithm to process pending shipments and calculate the best route
     *
     * @return Successful operation (status code 200)
     */
    @Operation(
        operationId = "processPendingShipments",
        summary = "Endpoint to process pending shipments",
        description = "Dijkstra algorithm to process pending shipments and calculate the best route",
        tags = { "Shipments" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Successful operation")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = ShipmentsApi.PATH_PROCESS_PENDING_SHIPMENTS
    )
    ResponseEntity<Void> processPendingShipments(
        
    );


    String PATH_PURGE_SHIPMENTS = "/shipments";
    /**
     * DELETE /shipments : Endpoint to purge shipments just processed (status DISCARDED or PLANNED)
     * Endpoint to purge shipments just processed (status DISCARDED or PLANNED). PENDING shipments will not be deleted
     *
     * @return Successful operation (status code 200)
     *         or The specified resource was not found (status code 404)
     */
    @Operation(
        operationId = "purgeShipments",
        summary = "Endpoint to purge shipments just processed (status DISCARDED or PLANNED)",
        description = "Endpoint to purge shipments just processed (status DISCARDED or PLANNED). PENDING shipments will not be deleted",
        tags = { "Shipments" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Successful operation"),
            @ApiResponse(responseCode = "404", description = "The specified resource was not found")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = ShipmentsApi.PATH_PURGE_SHIPMENTS
    )
    ResponseEntity<Void> purgeShipments(
        
    );

}